+++ apps/chunithm-scraper/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 For changelog before V4: [here](https://github.com/leomotors/chunithm-net-scraper/blob/main/CHANGELOG.md)
 
+## [4.0.0-rc.0] - 2025-04-17
+
+- Scrape function done and complete
+- todo: generate image
+
 ## [4.0.0-beta.0] - 2025-04-17
 
 - almost done